more learning, casting things to int

+++ CVstream.py
@@ -112,15 +112,15 @@ class CVStream(Process):
            Generate a proper mask from the set of proportional coordinates passed in.
            This gets called once as a setup function.
         """
-        self.mask = np.zeros((self.CAPTURE_W, self.CAPTURE_H))
+        self.mask = np.zeros((self.CAPTURE_W, self.CAPTURE_H), dtype=np.uint8)
         nonrels = []
         if len(self.settings.maskc) > 3:
             for relative_coordinate in self.settings.maskc:
                 nonrels.append(
-                    [relative_coordinate[0]*self.CAPTURE_W,
-                     relative_coordinate[1]*self.CAPTURE_H]
+                    [int(relative_coordinate[0]*self.CAPTURE_W),
+                     int(relative_coordinate[1]*self.CAPTURE_H)]
                     )
-            mask_points = np.array([nonrels], dtype=np.int32)
+            mask_points = np.array([nonrels], dtype=np.uint8)
             cv2.fillConvexPoly(self.mask, mask_points, 1)
             return True
         else:
